Split test into two files.
